+++ client/rpc/src/main/kotlin/net/corda/client/rpc/internal/RPCClientProxyHandler.kt
@@ -14,6 +14,9 @@ import com.google.common.util.concurrent.ThreadFactoryBuilder
 import net.corda.core.ThreadBox
 import net.corda.core.crypto.random63BitValue
 import net.corda.core.getOrThrow
+import net.corda.core.internal.LazyPool
+import net.corda.core.internal.LazyStickyPool
+import net.corda.core.internal.LifeCycle
 import net.corda.core.messaging.RPCOps
 import net.corda.core.serialization.KryoPoolWithContext
 import net.corda.core.utilities.*
