+++ core/src/main/kotlin/net/corda/core/internal/LazyPool.kt
@@ -1,7 +1,6 @@
-package net.corda.core.utilities
+package net.corda.core.internal
 
 import java.util.concurrent.ConcurrentLinkedQueue
-import java.util.concurrent.LinkedBlockingQueue
 import java.util.concurrent.Semaphore
 
 /**
