+++ node/src/main/kotlin/net/corda/node/shell/InteractiveShell.kt
@@ -18,7 +18,7 @@ import net.corda.core.internal.div
 import net.corda.core.internal.write
 import net.corda.core.messaging.CordaRPCOps
 import net.corda.core.messaging.StateMachineUpdate
-import net.corda.core.utilities.Emoji
+import net.corda.core.internal.Emoji
 import net.corda.core.utilities.loggerFor
 import net.corda.jackson.JacksonSupport
 import net.corda.jackson.StringToMethodCallParser
