+++ test-utils/src/main/kotlin/net/corda/testing/performance/Injectors.kt
@@ -3,7 +3,6 @@ package net.corda.testing.performance
 import com.codahale.metrics.Gauge
 import com.codahale.metrics.MetricRegistry
 import com.google.common.base.Stopwatch
-import net.corda.core.utilities.Rate
 import net.corda.testing.driver.ShutdownManager
 import java.time.Duration
 import java.util.*
